% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impGenotypeMatrix.R
\name{impGenotypeMatrix}
\alias{impGenotypeMatrix}
\title{Imputed missing genotypes in incomplete genotype matrix}
\usage{
impGenotypeMatrix(genotype_matrix)
}
\arguments{
\item{genotype_matrix}{Genotype matrix (sample x snps) coded with 0,1,2, or NA}
}
\description{
Returns a matrix with imputed missing data in a 0,1,2 coded genotype matrix
}
