% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImpPCA.R
\name{pca_plot}
\alias{pca_plot}
\title{Plot 1 PCA plot}
\usage{
pca_plot(loads, pca_x, pca_y, meta, pc_percent)
}
\arguments{
\item{loads}{matrix of PC loads}

\item{pca_x}{PC nr for x-axis}

\item{pca_y}{PC nr for y-axis}

\item{meta}{metadata for size and legend}

\item{pc_percent}{proportion of variance explained for all PCs}
}
\value{
PCA plot
}
\description{
Plot 1 PCA plot
}
