% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImpPCA.R
\name{plotImpPCA}
\alias{plotImpPCA}
\title{Plot PCAs 1-6}
\usage{
plotImpPCA(pca_obj, original_matrix, meta_file, output_pca_pdf)
}
\arguments{
\item{pca_obj}{PCA object from R funtion prcomp()}

\item{original_matrix}{Original genotype matrix coded 0,1,2, or <NA>}

\item{meta_file}{CSV file with two columns: sample-ID and population. Might contain additional columns.}

\item{output_pca_pdf}{Output file with ".pdf"}
}
\description{
Exports plots of PCs 1~2, 3~4, 5~6 to PDF file
}
