% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImpPCA.R
\name{plotImpPCA}
\alias{plotImpPCA}
\title{Plot PCAs 1-6}
\usage{
plotImpPCA(pca_gt, gt_imputed, gt, max_missing_snp = 1,
  max_missing_sample = 1, meta_file, output_pca_filename)
}
\arguments{
\item{max_missing_snp}{Set a max. percentage of missingness per SNP allowed. Default is set to 1 (i.e. 100 percent)}

\item{max_missing_sample}{Set a max. percentage of missingness per sample allowed.  Default is set to 1 (i.e. 100 percent)}

\item{meta_file}{CVS file that contains headers "Sample_ID" and "Population". Can contain other}
}
\description{
Exports plots of PCs 1~2, 3~4, 5~6 to PDF file
}
